corrected and updated orphan module
